adding plugin and renaming package to gormv2
