+++ module/apmgormv2/dialects/postgres/init.go
@@ -1,0 +1,26 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+// Package apmgormv2postgres imports the gorm postgres dialect package,
+// and also registers the lib/pq driver with apmsql.
+package apmgormv2postgres
+
+import (
+	_ "gorm.io/driver/postgres" // import the postgres dialect
+
+	_ "go.elastic.co/apm/module/apmsql/pq" // register lib/pq with apmsql
+)
