+++ module/apmgormv2/dialects/sqlite/init.go
@@ -15,9 +15,9 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// Package apmgormsqlite imports the gorm sqlite dialect package,
+// Package apmgormv2sqlite imports the gorm sqlite dialect package,
 // and also registers the sqlite3 driver with apmsql.
-package apmgormsqlite
+package apmgormv2sqlite
 
 import (
 	_ "gorm.io/driver/sqlite" // import the sqlite dialect
